Fixed some portability issues

+++ Emulator/Utilities/Types.h
@@ -39,14 +39,14 @@ typedef signed char        i8;
 typedef signed short       i16;
 typedef signed int         i32;
 typedef signed long long   i64;
-typedef ssize_t            isize;
+typedef signed long        isize;
 
 // Unsigned integers
 typedef unsigned char      u8;
 typedef unsigned short     u16;
 typedef unsigned int       u32;
 typedef unsigned long long u64;
-typedef size_t             usize;
+typedef unsigned long      usize;
 
 // Signed alternative for the sizeof keyword
 #define isizeof(x) (isize)(sizeof(x))
